==> Stage_3-DownloadAndPreparePosts.ipynb <==
# Extract Book Titles from The Posts

For every post id: get the actual version of the post, normalize it and search for book titles.

import datetime as dt
import multiprocessing
import re
from functools import partial
from pathlib import Path

import pandas as pd
import praw
import yaml
from tqdm.contrib.concurrent import process_map

from Functions import download_and_prepare_posts

NON_ALPHANUMERIC = re.compile(r'[\W_]+')
EXTRACTED_BOOKS_PATH = './data/books.csv'
SUBREDDIT_NAME = 'books'
IDS_PATH = './data/post_ids.csv'
STORED_POSTS_DF = './data/stored_posts.csv'

%load_ext autotime

### Connect  to API

with open('./keys', 'r') as file:
    keys = yaml.safe_load(file)

reddit = praw.Reddit(client_id=keys['client_id'],
                     client_secret=keys['client_secret'],
                     user_agent=keys['user_agent'])

list(reddit.subreddit('books').hot())[0].title

### Prepare Post IDs

post_ids_df = pd.read_csv(IDS_PATH, index_col=0)
post_ids_df

all_post_ids = set(zip(post_ids_df.ID, [dt.datetime.fromtimestamp(ts).year for ts in post_ids_df.Timestamp]))
del post_ids_df

### Prepare Existing Downloaded Posts

if Path(STORED_POSTS_DF).is_file():
    stored_posts_df = pd.read_csv(STORED_POSTS_DF, index_col=0)[['ID', 'Year', 'Post Text', 'Comment Text']]
else:
    stored_posts_df = pd.DataFrame(columns=['ID', 'Year', 'Post Text', 'Comment Text'])

stored_posts_ids = set(zip(stored_posts_df.ID, stored_posts_df.Year))
print(f'Loaded {len(stored_posts_ids):,} stored posts.')
stored_posts_df

all_post_ids = list(all_post_ids - stored_posts_ids)
print(f'{len(all_post_ids):,} left.')

### Download Posts and Comments and Normalize Them

# https://www.reddit.com/r/redditdev/comments/rsz7za/getting_submissions_from_praw_extremely_slow/
# https://www.reddit.com/r/redditdev/comments/atrt4i/praw_is_there_a_way_to_get_an_objects/

# retrieved_posts = set()
#
# fullnames = [f't3_{p_id}' for p_id, year in all_post_ids]
# deleted = {'[removed]', '[deleted]', '[deleted by user]'}
#
# for submission, (p_id, year) in (pbar := tqdm(zip(reddit.info(fullnames=fullnames), all_post_ids), total=len(all_post_ids), unit='post(s)')):
#     pbar.set_description(f'Process post with id: {p_id}', refresh=True)
#     title = submission.title
#     post_text = submission.selftext
#
#     if post_text not in deleted:
#         comments_fullnames = [f't3_{c_id}' for c_id in submission.comments.list()]
#         all_comments = [text for comment in reddit.info(fullnames=comments_fullnames) if (text := comment.selftext) not in deleted]
#         comments_text = normalize_text(''.join(all_comments))
#
#         retrieved_posts.add((p_id, year, normalize_text(title + post_text), comments_text))
#
# print(retrieved_posts)

def split(list_to_split, number_of_resulting_parts):
    k, m = divmod(len(list_to_split), number_of_resulting_parts)
    return [list_to_split[i*k+min(i, m):(i+1)*k+min(i+1, m)] for i in range(number_of_resulting_parts)]

# Split into chunks with size=100. This maximizes throughput for the reddit api.
all_post_ids_splitted = [all_post_ids[i:i+100] for i in range(0, len(all_post_ids), 100)]

results = process_map(
    partial(download_and_prepare_posts, praw_client_id=keys['client_id'], praw_client_secret=keys['client_secret'], praw_client_ua=keys['user_agent']),
    all_post_ids_splitted,
    max_workers=multiprocessing.cpu_count(),
)

print(f'Downloaded {len(results)} posts. Discarded {len(all_post_ids) - len(results)} posts.')

retrieved_posts = set().union(*results)

result_df = pd.DataFrame(retrieved_posts, columns=['ID', 'Year', 'Post Text', 'Comment Text'])
result_df

new_stored = pd.concat([stored_posts_df, result_df], ignore_index=True)
new_stored.replace(pd.NA, '', inplace=True)
new_stored.sort_values('Year', inplace=True, ignore_index=True)
new_stored.reset_index(inplace=True)
new_stored = new_stored[['ID', 'Year', 'Post Text', 'Comment Text']]
new_stored

new_stored.to_csv(STORED_POSTS_DF, encoding='utf-8')

print(f'Stored {len(new_stored):,} posts')
